// apps/server/functions/src/app/graphql.schema.interface.ts
export interface CompanyProfile {
    __typename?: 'CompanyProfile';
    id: string;
    compnayName: string;
    mainPhone: string;
    mainEmail: string;
    accountId: string;
    serviceCategory?: Nullable<string>;
    address?: Nullable<Address>;
    logoURL?: Nullable<string>;
    fax?: Nullable<string>;
    socials?: Nullable<Nullable<string>[]>;
    website?: Nullable<string>;
    created?: Nullable<string>;
    bio?: Nullable<string>;
}
// ...
export interface Address {
    __typename?: 'Address';
    streetAddress: string;
    city: string;
    state: string;
    zipcode: string;
}
// ...
type Nullable<T> = T | null;